prj/M2_Encoder/requirements.txt
torch==2.0.0
pytorch_lightning==1.9.2
transformers==4.17
Pillow
tqdm
ipdb==0.13.7
einops==0.3.0
pyarrow
sacred==0.8.2
pandas
timm==0.4.12
torchmetrics
torchvision
torchaudio
fairscale==0.4.1
protobuf==3.15.1
numpy
scipy
opencv-python
ftfy
sentencepiece
pyomo
mup
openspg-nn4k
modelscope

